requirements.txt
pygame
numpy
opencv-python
tobii-research
keyboard
os
PIL
pyyaml